Move Activation and User models into h.models package
