+++ h/accounts/views.py
@@ -17,8 +17,6 @@ from h import mailer
 from h import models
 from h import session
 from h.accounts import schemas
-from h.accounts.models import User
-from h.accounts.models import Activation
 from h.accounts.events import ActivationEvent
 from h.accounts.events import PasswordResetEvent
 from h.accounts.events import LogoutEvent
@@ -408,7 +406,7 @@ class ActivateController(object):
         except ValueError:
             raise httpexceptions.HTTPNotFound()
 
-        activation = Activation.get_by_code(self.request.db, code)
+        activation = models.Activation.get_by_code(self.request.db, code)
         if activation is None:
             self.request.session.flash(jinja2.Markup(_(
                 "We didn't recognize that activation link. "
@@ -420,7 +418,7 @@ class ActivateController(object):
             return httpexceptions.HTTPFound(
                 location=self.request.route_url('index'))
 
-        user = User.get_by_activation(self.request.db, activation)
+        user = models.User.get_by_activation(self.request.db, activation)
         if user is None or user.id != id_:
             raise httpexceptions.HTTPNotFound()
