+++ h/models/__init__.py
@@ -21,7 +21,6 @@ key to. So for convenience the test module can instead just do
 import sqlalchemy as sa
 from h.db import Base
 
-from h.accounts import models as accounts_models
 from h.auth import models as auth_models
 from memex.models.annotation import Annotation
 from memex.models.document import Document, DocumentMeta, DocumentURI
@@ -30,6 +29,9 @@ from h.groups import models as groups_models
 from h.nipsa import models as nipsa_models
 from h.notification import models as notification_models
 from h.badge import models as badge_models
+
+from h.models.activation import Activation
+from h.models.user import User
 
 __all__ = (
     'Activation',
@@ -47,7 +49,6 @@ __all__ = (
 )
 
 
-Activation = accounts_models.Activation
 Blocklist = badge_models.Blocklist
 Feature = features_models.Feature
 FeatureCohort = features_models.FeatureCohort
@@ -55,7 +56,6 @@ Group = groups_models.Group
 NipsaUser = nipsa_models.NipsaUser
 Token = auth_models.Token
 Subscriptions = notification_models.Subscriptions
-User = accounts_models.User
 
 
 def includeme(_):
